+++ lib/plugins/__init__.py
@@ -1,1 +1,0 @@
-from NeoVintageous.lib.plugins.plugins import register # noqa